requirements.txt
contourpy>=1.2,<2.0
cycler>=0.12,<1.0
filelock>=3.14,<4.0
fonttools>=4.53,<5.0
fsspec>=2024.6,<2025.0
graphviz>=0.19,<1.0
ImageHash>=4.3,<5.0
Jinja2>=3.1,<4.0
kiwisolver>=1.4,<2.0
MarkupSafe>=2.1,<3.0
matplotlib>=3.8,<4.0
mpmath>=1.3,<2.0
networkx>=3.3,<4.0
numpy>=1.26,<2.0
packaging>=24.0,<25.0
pillow>=10.0,<11.0
pyparsing>=3.1,<4.0
python-dateutil>=2.9,<3.0
PyWavelets>=1.6,<2.0
scipy>=1.13,<2.0
six>=1.16,<2.0
sympy>=1.12,<2.0
torch>=2.0,<3.0
torchaudio>=2.0,<3.0
torchvision>=0.15,<1.0
torchviz>=0.0.2,<0.1
triton>=2.0,<3.0
typing_extensions>=4.12,<5.0
